Beginning of package structured.
The structured package defines a minimal abstract structured storage
API for use by Filecoin actors.

This isolates Filecoin from the implementation details of that
interface and reduces accidental dependence, ensuring we can change
the implementation in the future should we desire.

It also allows provides a place to adapt the implementation's
interface to exactly what we want for our use case.

The API is not functional yet. The main thing needed to complete it
is to add Noms support to Cid.