Replace perform real proofs option with env var
We replace the confusingly named "perform-real-proofs" option for the
daemon with a lighter weight FILECOIN_USE_SMALL_SECTORS environment
variable that does the same thing. The old configuration option was
confusingly named and also odd to have a configuration option, as it
should be part of the blockchain configuration itself.

The environment variable solution is designed to be temporary in hopes
that we can the sector size to the genesis block in the future.